feat: formalize and rename progress bar component
