+++ src/main.ts
@@ -1,5 +1,5 @@
 export { PrimaryCard } from "./components/cards/PrimaryCard/PrimaryCard"
-export { ProgressBar } from "./components/cards/ProgressBar/ProgressBar"
+export { StepProgress } from "./components/charts/StepProgress/StepProgress"
 export { SecondaryCard } from "./components/cards/SecondaryCard/SecondaryCard"
 export { Avatar } from "./components/content/Avatar/Avatar"
 export { Icon } from "./components/content/Icon/Icon"
